Calculation Game
This is a simple Calculation Game written in PYTHON using only basic programming concepts like if, elif, and else statements. The game asks the user three questions and checks their answers. Based on the number of correct answers, it gives a score and feedback. It's a fun way to practice basic math and learn simple decision-making in PYTHON.